requirements.txt
Django==3.0.6
django-cart
stripe